refactor: `FormElement` 컴포넌트를 하위 디렉토리로 이동
