+++ src/entities/form/ui/DoubleSliderFormElement/DoubleSliderFormElement.tsx
