+++ src/entities/form/ui/SelectorFormElement/SelectorFormElement.tsx
