+++ src/entities/form/ui/SliderFormElement/SliderFormElement.stories.tsx
@@ -1,29 +1,28 @@
 import { FormStateContextProvider } from "@/entities/form/contexts";
 import { FormSchema } from "@/entities/form/types";
-import { SelectorFormElement } from "@/entities/form/ui/SelectorFormElement";
+import { SliderFormElement } from "@/entities/form/ui/SliderFormElement/SliderFormElement";
 import type { Meta, StoryObj } from "@storybook/react";
 
-const meta: Meta<typeof SelectorFormElement> = {
-    title: "Entity/Form/SelectorFormElement",
-    component: SelectorFormElement,
+const meta: Meta<typeof SliderFormElement> = {
+    title: "Entity/Form/SliderFormElement",
+    component: SliderFormElement,
 };
 
 export default meta;
-type Story = StoryObj<typeof SelectorFormElement>;
+type Story = StoryObj<typeof SliderFormElement>;
 
 const formInitialState: FormSchema = {
     id: 1,
-    title: "선호하는 색상",
-    description: "선호하는 색상을 선택해주세요.",
+    title: "잠귀",
+    description: "잠귀에 대한 민감도를 선택해주세요.",
     questions: [
         {
             id: 1,
-            title: "선호하는 색상이 무엇인가요?",
-            type: "SELECTOR",
+            title: "잠귀에 대한 민감도를 선택해주세요.",
+            type: "SLIDER",
             options: [
-                { label: "빨강", value: "Red" },
-                { label: "초록", value: "Green" },
-                { label: "파랑", value: "Blue" },
+                { label: "밝음", value: "1" },
+                { label: "어두움", value: "10" },
             ],
             optionDelimiter: null,
         },
@@ -37,13 +36,13 @@ export const Default: Story = {
     render: (args) => {
         return (
             <FormStateContextProvider formInitialState={formInitialState}>
-                <SelectorFormElement
+                <SliderFormElement
                     id={args.id}
                     title={args.title}
                     options={args.options}
-                    type={args.type}
                     optionDelimiter={args.optionDelimiter}
-                />
+                    type={args.type}
+                ></SliderFormElement>
             </FormStateContextProvider>
         );
     },
